fix: invalid write of dropped promise
A future never suspending drops its promise when returning to the caller,
which means when executing the destructor, its promise is already dropped.
But the outermost future has no awaiter,
so its promise can only be dropped by completing the coroutine
and we can ensure the destructor always executed before completing the coroutine.