[Fix] Reverts #41 (#74)
# Fix
- Reverts #41 due to the corner case in multi-thread runtime:
1. A chained coroutine reaches the suspend point. It registers a event
to the driver and then the underlying thread was suspended by the os.
2. The event in the driver is ready and another thread resumes the
coroutine. It runs till the end and the underlying objects are
destructed.
3. First thread resumes, so it continues at the first line after the
`resume()` in `await_suspend`. But the next blocks contains references
to the destructed `Future` which leads to ub.
    
Cons: In a nested no suspend coroutine(like coroutine recursion or
coroutine in a loop), the stack depth grows twice as the speed of the
usual recursion.
- Syncs latest llvm toolchain installation script from
https://apt.llvm.org/llvm.sh
- Fixes bazel coverage outputs empty coverage data for ut