[Fix] catch poll wrapper exception (#110)
- Bug
If some of the `select` and `join` tasks throw uncaught exceptions, the
exceptions unwind directly to the `run_loop_once` instead of through the
coroutine chain.

- Root Cause
When the suspended `Future` is resumed and the `poll` method throws an
exception, the stack looks like: `run_loop_once` -> `poll_wrapper` ->
`poll`. So the behavior here complies with the exception handling in
synchronous code.

- Solution
Align `Future` exception unwinding behavior with `unhandled_exception`
of coroutine promises, that's storing a `expcetion_ptr` and resumes the
dependent coroutine normally.